rename: query 관련 파일을 query 폴더로 이동
