+++ src/service/query/member.ts
@@ -1,4 +1,4 @@
-import { client } from './sanity';
+import { client } from '../sanity';
 
 export async function addMember(uid: string) {
   return client.createIfNotExists({
